requirements.txt
streamlit==1.43.2
google-generativeai
openpyxl==3.1.5
pymupdf==1.25.3
python-pptx==1.0.2
pytest==8.3.5
pytest-mock==3.14.0
